requirements.txt
requests
psycopg2
qdrant-client
python-dotenv
python-dateutil